requirements.txt
asgiref==3.2.7
attrs==19.3.0
aws-sam-translator==1.23.0
aws-xray-sdk==2.5.0
boto==2.49.0
boto3==1.13.2
botocore==1.16.2
certifi==2020.4.5.1
cffi==1.14.0
cfn-lint==0.30.1
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-cleanup==4.0.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-heroku==0.3.1
django-hitcount==1.3.2
django-storages==1.9.1
django-taggit==1.2.0
django-widget-tweaks==1.4.8
docker==4.2.0
docutils==0.15.2
ecdsa==0.15
future==0.18.2
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.6.0
importlib-resources==1.5.0
Jinja2==2.11.2
jmespath==0.9.5
jsondiff==1.1.2
jsonpatch==1.25
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
MarkupSafe==1.1.1
mock==4.0.2
moto==1.3.14
networkx==2.4
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-jose==3.1.0
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
responses==0.10.14
rsa==4.0
s3transfer==0.3.3
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
sshpubkeys==3.1.0
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.1
whitenoise==5.0.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.1.0
